+++ Sources/Soprano/Controllers/KeybindingManager.swift
@@ -42,6 +42,7 @@ struct PaneNavigationClaimRegistry {
 protocol KeybindingDelegate: AnyObject {
     func keybindingToggleSidebar()
     func keybindingSaveSession()
+    func keybindingSwitchToLastSession()
     func keybindingRenameWindow()
     func keybindingOpenSettings()
     func keybindingOpenAgentDashboard()
@@ -451,6 +452,8 @@ final class KeybindingManager: @unchecked Sendable {
             agentManager.activateNextWindow()
         case "last-window":
             agentManager.activateLastWindow()
+        case "last-session":
+            invokeDelegate { $0.keybindingSwitchToLastSession() }
 
         case "resize-left":
             agentManager.resizePane(direction: .left, tickPercent: config.resizeTickPercent)